requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.9.2
python-dotenv==1.0.1
alembic==1.13.2